feat(waf): add 6 new WAF vendors, bypass techniques, and extract signatures
