+++ packages/core/src/waf-signatures.ts
@@ -1,0 +1,324 @@
+export interface WAFSignature {
+	name: string;
+	headers: { [key: string]: string | RegExp };
+	statusCodes?: number[];
+	bodyPatterns?: RegExp[];
+	cookiePatterns?: RegExp[];
+	responseTime?: {
+		min?: number;
+		max?: number;
+	};
+}
+
+export const WAF_SIGNATURES: WAFSignature[] = [
+	// Cloudflare
+	{
+		name: 'Cloudflare',
+		headers: {
+			server: /cloudflare/i,
+			'cf-ray': /^[a-f0-9]+-[A-Z]{3}$/,
+			'cf-cache-status': /.*/,
+			'cf-request-id': /.*/,
+			'cf-mitigated': /.*/,
+		},
+		statusCodes: [403, 429],
+		bodyPatterns: [/cloudflare/i, /attention required! \| cloudflare/i, /ray id: [a-f0-9]+-[A-Z]{3}/i, /Cloudflare Ray ID:/i],
+	},
+
+	// AWS WAF
+	{
+		name: 'AWS WAF',
+		headers: {
+			server: /CloudFront|awselb/i,
+			'x-amz-cf-id': /.*/,
+			'x-amz-cf-pop': /.*/,
+			'x-cache': /^(Hit|Miss) from cloudfront$/i,
+			'x-amzn-requestid': /.*/,
+		},
+		statusCodes: [403],
+		bodyPatterns: [/forbidden.*you don't have permission to access.*on this server/i, /request blocked/i, /Request blocked\./i],
+	},
+
+	// Imperva/Incapsula
+	{
+		name: 'Imperva',
+		headers: {
+			'x-iinfo': /.*/,
+			'x-cdn': /Incapsula/i,
+			'set-cookie': /incap_ses_|visid_incap_/i,
+		},
+		statusCodes: [403],
+		bodyPatterns: [/incapsula/i, /request unsuccessful. incapsula incident id/i, /generated by cloudflare/i],
+		cookiePatterns: [/incap_ses_\d+/, /visid_incap_\d+/],
+	},
+
+	// F5 BIG-IP
+	{
+		name: 'F5 BIG-IP',
+		headers: {
+			server: /BIG-IP/i,
+			'x-wa-info': /.*/,
+			'f5-trace-id': /.*/,
+		},
+		statusCodes: [403],
+		bodyPatterns: [/the requested url was rejected/i, /please consult with your administrator/i, /your support id is/i],
+	},
+
+	// ModSecurity
+	{
+		name: 'ModSecurity',
+		headers: {
+			server: /mod_security|apache/i,
+		},
+		statusCodes: [403, 406],
+		bodyPatterns: [/mod_security/i, /not acceptable/i, /apache.*forbidden/i, /request blocked by security policy/i],
+	},
+
+	// Akamai
+	{
+		name: 'Akamai',
+		headers: {
+			server: /AkamaiGHost/i,
+			'akamai-origin-hop': /.*/,
+			'x-akamai-transformed': /.*/,
+			'x-akamai-request-id': /.*/,
+		},
+		statusCodes: [403],
+		bodyPatterns: [/access denied/i, /akamai/i, /reference #[0-9a-f]+/i],
+	},
+
+	// Barracuda
+	{
+		name: 'Barracuda',
+		headers: {
+			server: /Barracuda/i,
+			'x-barracuda-url': /.*/,
+		},
+		statusCodes: [403],
+		bodyPatterns: [/barracuda/i, /access denied/i],
+	},
+
+	// Sucuri
+	{
+		name: 'Sucuri',
+		headers: {
+			server: /Sucuri/i,
+			'x-sucuri-id': /.*/,
+			'x-sucuri-cache': /.*/,
+		},
+		statusCodes: [403],
+		bodyPatterns: [/sucuri website firewall - access denied/i, /questions\? contact us at cloudproxy@sucuri\.net/i],
+	},
+
+	// Fastly
+	{
+		name: 'Fastly',
+		headers: {
+			via: /fastly/i,
+			'x-served-by': /cache-.*-fastly/i,
+			'x-cache': /(HIT|MISS).*fastly/i,
+		},
+		statusCodes: [403],
+	},
+
+	// KeyCDN
+	{
+		name: 'KeyCDN',
+		headers: {
+			server: /keycdn-engine/i,
+			'x-edge-location': /.*/,
+		},
+		statusCodes: [403],
+	},
+
+	// StackPath (MaxCDN)
+	{
+		name: 'StackPath',
+		headers: {
+			server: /NetDNA-cache|stackpath/i,
+			'x-hw': /.*/,
+		},
+		statusCodes: [403],
+	},
+
+	// DenyAll
+	{
+		name: 'DenyAll',
+		headers: {
+			server: /denyall/i,
+		},
+		statusCodes: [403],
+		bodyPatterns: [/denyall/i],
+	},
+
+	// Fortinet FortiWeb
+	{
+		name: 'FortiWeb',
+		headers: {
+			server: /Fortigate|FortiWeb/i,
+		},
+		statusCodes: [403],
+		bodyPatterns: [/web filter violation/i, /fortigate/i],
+	},
+
+	// Wallarm
+	{
+		name: 'Wallarm',
+		headers: {
+			server: /nginx-wallarm/i,
+			'x-wallarm-instance': /.*/,
+		},
+		statusCodes: [403, 500],
+	},
+
+	// Radware AppWall
+	{
+		name: 'Radware',
+		headers: {
+			server: /Radware|AppWall/i,
+			'x-origin-requestid': /.*/,
+		},
+		statusCodes: [403],
+	},
+
+	// Azure Front Door
+	{
+		name: 'Azure Front Door',
+		headers: {
+			'x-azure-ref': /.*/,
+			server: /Microsoft-HTTPAPI\/2\.0/i,
+		},
+		statusCodes: [403],
+		bodyPatterns: [/Our services aren't available right now/i, /Your request has been blocked/i, /Microsoft-Azure-Application-Gateway/i],
+	},
+
+	// Google Cloud Armor
+	{
+		name: 'Google Cloud Armor',
+		headers: {
+			server: /^GSE$/i,
+		},
+		statusCodes: [403, 404],
+		bodyPatterns: [/Request blocked by Cloud Armor/i, /Access Denied.*Cloud Armor/i],
+	},
+
+	// Citrix NetScaler
+	{
+		name: 'Citrix NetScaler',
+		headers: {
+			server: /NetScaler/i,
+			'vi-id': /.*/,
+		},
+		cookiePatterns: [/ns_af=/i, /citrix_ns_id/i],
+		statusCodes: [403],
+		bodyPatterns: [/The requested URL was rejected\. Please consult with your administrator/i, /NS-CACHE/i],
+	},
+
+	// Varnish (often used with WAF modules)
+	{
+		name: 'Varnish',
+		headers: {
+			server: /varnish/i,
+			'x-varnish': /.*/,
+			via: /varnish/i,
+		},
+		statusCodes: [403],
+	},
+
+	// Palo Alto Networks
+	{
+		name: 'Palo Alto Networks',
+		headers: {
+			'x-phx': /.*/,
+		},
+		statusCodes: [403],
+		bodyPatterns: [/Virus\/Spyware Download Blocked/i, /Palo Alto Next Generation Security Platform/i, /Access Denied/i],
+	},
+
+	// Sophos WAF
+	{
+		name: 'Sophos WAF',
+		headers: {
+			'x-sophos-waf-id': /.*/,
+		},
+		cookiePatterns: [/sophos_waf_id/i],
+		statusCodes: [403],
+		bodyPatterns: [/Powered by UTM Web Protection/i, /Sophos Firewall/i],
+	},
+
+	// --- New Signatures Added ---
+	
+	// Signal Sciences (Fastly Next-Gen WAF)
+	{
+		name: 'Signal Sciences',
+		headers: {
+			'x-sigsci-request-id': /.*/,
+			'x-sigsci-tags': /.*/,
+		},
+		statusCodes: [406],
+		bodyPatterns: [/Not Acceptable/i],
+	},
+
+	// NGINX App Protect / NAXSI
+	{
+		name: 'NGINX App Protect',
+		headers: {
+			server: /nginx/i,
+		},
+		statusCodes: [403],
+		bodyPatterns: [/The request was rejected/i, /NAXSI/i, /Request Rejected/i],
+	},
+
+	// HAProxy Enterprise WAF
+	{
+		name: 'HAProxy',
+		headers: {},
+		statusCodes: [403],
+		bodyPatterns: [/HAProxy/i, /Protected by HAProxy/i, /Request Blocked/i],
+	},
+
+	// IBM DataPower
+	{
+		name: 'IBM DataPower',
+		headers: {
+			'x-backside-transport': /.*/,
+			'x-dp-watson-tran-id': /.*/,
+		},
+		statusCodes: [403],
+		bodyPatterns: [/DataPower/i, /Transaction Error/i, /Access Denied/i],
+	},
+
+	// Reblaze
+	{
+		name: 'Reblaze',
+		headers: {
+			server: /Reblaze Secure Web Gateway/i,
+		},
+		cookiePatterns: [/rbzid=/i],
+		statusCodes: [403],
+		bodyPatterns: [/Reblaze/i, /Access denied by Reblaze/i],
+	},
+
+	// DotDefender
+	{
+		name: 'DotDefender',
+		headers: {
+			'x-dotdefender-denied': /.*/,
+		},
+		statusCodes: [403],
+		bodyPatterns: [/dotDefender Blocked Your Request/i],
+	},
+
+	// Generic detection patterns — require WAF-specific phrases,
+	// not just common HTTP words like "forbidden"
+	{
+		name: 'Generic WAF',
+		headers: {},
+		statusCodes: [403, 406, 429],
+		bodyPatterns: [
+			/request.*(blocked|rejected|filtered).*by/i,
+			/web.*(firewall|application firewall|waf)/i,
+			/malicious.*request|suspicious.*activity|attack.*detected/i,
+		],
+	},
+];
